requirements.txt
pyrogram
httpx
tgcrypto
python-dotenv
